M-sized-integers step 7: a byte says byte, and a widening stops pretending it can fail — docs/debrief/QUEUE.md
`s[i]` is a `u8`. The author's rule was "the precise type, not the widest one
that holds it", and this is the language's only byte source — before it, `[u8]`
was describable and unbuildable, which panel 042's llm-ergonomist found by
writing the task the widths exist for and running out of language.

THE MIGRATION IS A CORRECTION, NOT A TAX, and that is the finding. Six programs
broke. Three of the repairs are these:

    function is_digit(c: i64) -> bool   ->   function is_digit(c: u8) -> bool
    function digit_of(c: i64) -> i64?   ->   function digit_of(c: u8) -> i64?
    function here(l: Lex) -> i64        ->   function here(l: Lex) -> u8

Three functions that took or returned BYTES and declared 64-bit integers. The
precise type did not weigh the programs down; it made approximate signatures
true. 05-mutation.hero needed one line.

AND THE MEASUREMENT OVERTURNED A CALL I MADE THREE STEPS AGO. The author ruled
that conversions return `T?`; I applied it uniformly and overruled the
llm-ergonomist, which had asked for widening to be infallible. Then `s[i] -> u8`
put the cost on the language's own acceptance programme:

    v @ v * 10 + fit_i64(l.here() - '0').must()

A `.must()` on a conversion that cannot fail, at the hottest line of a lexer —
the exact shape the closure list is made of. So `fit_<w>` now returns `T` where
the source width is provably contained in the target and `T?` where it is not.

This bends the author's ruling and is marked PENDING RATIFICATION, but it does
not bend panel 042's Q2, which refused one name with two result types on the
ground of UNPREDICTABILITY. This is predictable: the result is a function of the
argument's width, the compiler knows it, and it is fallible exactly when it can
fail. `IntKind::contains` is the one place that decides, and its four cases are
deliberately asymmetric — an unsigned source needs a STRICTLY wider signed
target, because the target spends a bit on a sign it will never use, and a
signed source fits no unsigned target however wide, because -1 is a value.

The index is NOT narrowed and that is deliberate: Rust accepts only `usize`
there, and the historian reported the `as usize` tax as the single
most-complained-of integer decision in its whole survey.

Independent confirmation that the widening is transparent: the mangler's hash,
computed in Heroes over `rendered[i]` and now through `fit_i64`, produces the
same four values it produced before — 452710609, 947353345, 1680613593,
323221308.

Spec 2959 -> 2956, MINUS three. "yields an `i64` in 0..255 (a byte)" was a type,
a range and a gloss where "yields a `u8`" is a type. The document got shorter by
following the author's own rule.

design.md's appendix migrated with the rest: it is compiled by the acceptance
tests, so it could not be left behind.

532 tests green, clippy clean.

+++ crates/heroes/src/types/tests/calls.rs
@@ -77,7 +77,7 @@ function main()
     // where nothing can be inverted.
     assert_clean(
         "\
-function at(text: str, index: i64) -> i64
+function at(text: str, index: i64) -> u8
     return text[index]
 
 function main()
